requirements.txt
# Google APIs and Authentication
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
uritemplate==4.2.0
urllib3==2.5.0

# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AI and ML
google-generativeai==0.3.2

# System Operations
psutil==5.9.6

# CLI and Terminal
click==8.1.7
rich==13.7.0

# Desktop & Web Automation
pyautogui==0.9.54
selenium==4.16.0
webdriver-manager==4.0.1
Pillow>=10.2.0
opencv-python>=4.9.0.80
pynput==1.7.6

# Development
python-dotenv==1.0.0

# GUI and Audio
matplotlib>=3.7.0
numpy>=1.24.0
pyaudio>=0.2.11

# Research and Reports
markdown>=3.5.0

# Memory and Learning
pygetwindow>=0.0.9

# Speech Recognition and Wake Word Detection
SpeechRecognition==3.14.3
audioop-lts==0.2.2
standard-aifc==3.13.0
standard-chunk==3.13.0

# Enhanced Text-to-Speech Services
edge-tts>=6.1.0
google-cloud-texttospeech>=2.16.0
pygame>=2.5.0
gtts>=2.4.0
pyttsx3>=2.90
